% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cap}
\alias{cap}
\title{Convert the first letter of a string to a capital}
\usage{
cap(x)
}
\arguments{
\item{x}{The string to be converted.}
}
\value{
The capitalized string.
}
\description{
If the first letter of \code{x} is a lowercase letter, this function converts it to a capital and returns
the resulting string. Otherwise the string \code{x} is returned.
}
