% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front_ends.R
\name{flip}
\alias{flip}
\title{Simulate coin flips followed by normal proportions inference}
\usage{
flip(
  n = 1,
  id = "fair",
  N = 1,
  confidence_level = NA,
  null_value = NA,
  true_value = NA,
  significance_level = 0.05,
  alternative = "two-sided",
  sync = NULL,
  sides = c("H", "T"),
  success = "H",
  show_summary = TRUE,
  show_plot = TRUE,
  xlim = NA,
  ylim = NA,
  digits = 6
)
}
\arguments{
\item{n}{A vector of one or two positive integers. A single positive integer
specifies the number of simulated flips of a coin per iteration. Two
positive integers specify the numbers of simulated flips for each of two
coins per iteration.}

\item{id}{The unique one or two id's of the coins whose flips are to be
simulated. These can be given either as vectors or lists. See the details
below.}

\item{N}{A positive integer specifying the number of iterations to simulate,
each iteration having the number of simulated flips specified by \code{n}.}

\item{confidence_level}{A number giving the confidence level of the
confidence intervals to be computed.}

\item{null_value}{A number giving the value of the test parameter under the
null hypothesis in the hypothesis tests to be conducted.}

\item{true_value}{A number giving what is to be taken as the true value of
the test parameter, for the purpose of determining which of the
computed confidence intervals contain the true value.}

\item{significance_level}{A number giving the significance level for the
hypothesis tests to be conducted, for the purpose of determining which
of the hypothesis tests attained statistical significance.}

\item{alternative}{A character string specifying which type of alternative
hypothesis should be used: either "two-sided" (or "two.sided"), "less",
or "greater".}

\item{sync}{An integer or character string that is used to set the random
seed immediately prior to the simulations. An integer or \code{NULL} will
be passed to \code{set.seed} directly, while a character string will
first be converted to an integer by \code{\link{make_seed}} and then
passed to \code{set.seed}.}

\item{sides}{This can be given either as a character vector or factor of
length 2, or as a list of two such character vectors or factors. A single
character vector or factor specifies the labels for the sides of a
coin that will be simulated; two such vectors or factors specify the
labels for two coins to be simulated.}

\item{success}{A character string identifying which label from \code{sides}
should be treated as "success" (in the sense of a Bernoulli trial) in
the simulated coin flips.}

\item{show_summary}{Boolean indicating: should a text summary be printed?}

\item{show_plot}{Boolean indicating: should a plot be displayed?}

\item{xlim}{Numeric vector of length 2 giving the lower and upper limits
of the x axis on the plot. NA means to use the default limits, if any.}

\item{ylim}{Numeric vector of length 2 giving the lower and upper limits
of the y axis on the plot. NA means to use the default limits, if any.}

\item{digits}{Integer giving the number of digits to round values to. A
value of NA indicates no rounding should be done.}
}
\value{
A ggplot2 plot object (whether or not such an object was
displayed).
}
\description{
\code{flip} simulates \code{N} iterations of \code{n} flips of a coin or two
}
\details{
This function simulates flipping one or two coins a specified number of
times (which constitutes a single "iteration"), and then repeating that
process again and again, for a total of \code{N} iterations. Then some
sort of descriptive of inferential statistics is applied to each
iteration, and the results of these are summarized in written and
graphical forms. DETAILS FROM HERE.
}
