% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fround}
\alias{fround}
\title{Round a number and convert the output to a string}
\usage{
fround(x, digits = 6, exact = TRUE)
}
\arguments{
\item{x}{The number to be rounded.}

\item{digits}{The number of decimal places to round to.}

\item{exact}{Boolean. \code{TRUE} specifies that the number of decimal places to round to is exact.}
}
\value{
String of the resulting rounded number.
}
\description{
This function rounds the number \code{x} to the number of digits specified in \code{digits}. If \code{exact} is
\code{TRUE}, then \code{digits} will be the exact number of decimal places (including any zeroes at the end).
Otherwise, trailing zeroes will be removed. The rounded number is then returned as a string (and not in
scientific notation).
}
