% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_in_order}
\alias{is_in_order}
\title{Test whether the arguments are in order from least to greatest}
\usage{
is_in_order(a, b, c = NA)
}
\arguments{
\item{a}{First number in the sequence whose order is to be tested.}

\item{b}{Second number in the sequence whose order is to be tested.}

\item{c}{Optional third number in the sequence whose order is to be tested.}
}
\value{
Boolean that is \code{TRUE} if the arguments are in nondescending order, \code{FALSE} if not.
}
\description{
If \code{c} is specified, then this function returns \code{TRUE} if \eqn{a \leq b \leq c} and \code{FALSE} otherwise.
If \code{c} is not specified, then this function returns \code{TRUE} if \eqn{a \leq b} and \code{FALSE} otherwise.
}
