% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_samplers.R
\name{iterate}
\alias{iterate}
\title{Repeat a sampling process multiple times}
\usage{
iterate(sampler, n = 1, N = 1)
}
\arguments{
\item{sampler}{A function of a single argument \code{n}. When called, this
function should generate a sample of size \code{n}.}

\item{n}{A positive integer giving the sample size for each sample that is
to be repeated.}

\item{N}{A positive integer giving the number of samples of size \code{n} to
produce.}
}
\value{
A list of length \code{N}.
}
\description{
Produce a list of samples from the specified sampling function
}
\details{
This function produces a list of length \code{N}, each of whose entries is a sample
generated by calling \code{sampler(n)}.
}
