% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join}
\alias{join}
\title{Convert a vector to a string}
\usage{
join(x, and = " and ", comma = ", ")
}
\arguments{
\item{x}{The vector to be converted to a string.}

\item{and}{The string to precede the last element in the vector for vectors of length at least 2.}

\item{comma}{The string to separate elements of the vector in the output string}
}
\value{
String listing the elements of the vector \code{x}.
}
\description{
This function converts a vector to a string consisting of the elements of the vector separated by
the string specified in \code{comma}, and with the last element preceded by the string
specified in \code{and}. Vectors of length 1 and 2 are also handled according to the usual
English language conventions for listing elements.
}
