% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_inferers.R
\name{list_infer}
\alias{list_infer}
\title{Apply an inference function to each element of a list}
\usage{
list_infer(x, inferer, ...)
}
\arguments{
\item{x}{A list of data, each element of which can be processed by the
function given as the \code{inferer} argument.}

\item{inferer}{A function whose first arguments (in order) are given by the
elements within each element of \code{x}, and which returns an S3 object
of class \code{htest}.}

\item{...}{Any remaining arguments to give to \code{inferer} beyond those
provided by each element of \code{x}.}
}
\value{
An object of class \code{htest_list}.
}
\description{
\code{list_infer} applies \code{inferer} to each element of \code{x}
}
\details{
The argument \code{x} should be a list, and the argument \code{inferer}
should be a function that returns an S3 object of class \code{htest}.
The \code{list_infer} function applies
\code{inferer} individually to each element of \code{x}. Because of
this, each element of \code{x} should either be suitable as the
first argument of \code{inferer}, or should itself be a list whose
elements are the first several arguments of \code{inferer}.

Any arguments given to \code{list_infer} in \code{...} will be provided
to \code{inferer} as further arguments, beyond those given within
each element of \code{x}.
}
