% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_samplers.R
\name{make_coin}
\alias{make_coin}
\title{Make a function that simulates coin flips}
\usage{
make_coin(id = "fair", sides = c("H", "T"))
}
\arguments{
\item{id}{A string or numeric vector, or a list of these, one for
each sample that the output function is to produce. As described in
\code{\link{make_finite_sampler}},
this is used to determine the probabilities for the \code{sides}.}

\item{sides}{A vector from which the output function is to
sample from, or a list of such vectors, one vector for each sample
that the output function is to produce. In essence, these vectors
give labels for the sides of the coins whose flips are to be simulated.}
}
\value{
This function outputs a function of a single argument \code{n}.
}
\description{
\code{make_coin} outputs a function that simulates coin flips.
}
\details{
This function is a convenience wrapper function for
\code{\link{make_finite_sampler}} with \code{id} as its first argument
and with \code{c("H","T")} as the default
value for \code{sides}.
}
