% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_samplers.R
\name{make_die}
\alias{make_die}
\title{Make a function that simulates die rolls}
\usage{
make_die(id = "fair", sides = 1:6)
}
\arguments{
\item{id}{A string or numeric vector, or a list of these, one for
each sample that the output function is to produce. As described in
\code{\link{make_finite_sampler}},
this is used to determine the probabilities for the \code{sides}.}

\item{sides}{A vector from which the output function is to
sample from, or a list of such vectors, one vector for each sample
that the output function is to produce. In essence, these vectors
give labels for the sides of the dice whose rolls are to be simulated.}
}
\value{
This function outputs a function of a single argument \code{n}.
}
\description{
\code{make_die} outputs a function that simulates die rolls.
}
\details{
This function is a convenience wrapper function for
\code{\link{make_finite_sampler}} with \code{id} as its first argument
and with \code{1:6} as the default
value for \code{sides}. In addition, while
\code{\link{make_finite_sampler}} allows all types of vectors for
its \code{sides} argument, \code{make_die} expects \code{sides} to
be a numeric vector (or a list of numeric vectors) and will coerce
it to such if need be.
}
