% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_samplers.R
\name{make_finite_sampler}
\alias{make_finite_sampler}
\title{Make a function that samples from a finite set.}
\usage{
make_finite_sampler(sides, id = "fair")
}
\arguments{
\item{sides}{A vector from which the output function is to
sample from, or a list of such vectors, one vector for each sample
that the output function is to produce.}

\item{id}{A string or numeric vector, or a list of these, one for
each sample that the output function is to produce. As described below,
this is used to determine the probabilities with which the \code{sides}
are to be sampled (with replacement).}
}
\value{
This function outputs a function of a single argument \code{n}.
}
\description{
\code{make_finite_sampler} returns a function that samples from a finite set.
}
\details{
In its simplest use, \code{make_finite_sampler} outputs a function of a
single argument \code{n} that generates
samples (with replacement) of size \code{n} from \code{sides}. This
output function can be thought of way to generalize simulating
repeated coin flips (where \code{sides = c("H","T")}) and die rolls
(where \code{sides = 1:6}).

The probabilities with which the output function samples the entries of
\code{sides} are determined by \code{id}, in three different ways:

\enumerate{
\item If \code{id} is \code{"fair"},
then the probabilities of all the sides will be equal.
\item If \code{id} is any other character string, the probabilities
are determined by a deliberately opaque method. This allows
\code{make_finite_sampler} to create sampling functions whose
probabilities are unknown and must be explored by statistical methods.
\item If \code{id} is a numeric vector with positive entries, that
vector will first be repeated or shortened until it has the same length
as \code{sides}. Then it will be rescaled so that its sum is 1. The
resulting numeric vector will give the probabilities with which the output
function will sample the entries of \code{sides}. This allows
\code{make_finite_sampler} to create sampling functions whose probabilities
are known. Such functions can be used to verify and assess statistical
methods.
}
}
\section{Multiple samples}{

The \code{make_finite_sampler} function can also be used to make a
function that produces more than one simulataneous
sample. In this case, the output function is a function of not just a
number \code{n}, but of a numeric vector \code{n}, which will be repeated
or shortened until its length matches the number of output samples
to be produced.

To create such a function, instead of giving the \code{id} argument
as a single string or single numeric vector, give it
as a list of single strings and/or single numeric vectors. (The list
can contain one or both types.) Each string or numeric vector
will correspond to a single sample produced by the output function.

For another way to create such a function, instead of giving the
\code{sides} argument as a single vector, give it as a list of vectors.
Each vector in the list will correspond to a single sample produced by
the output function.

If both \code{id} and \code{sides} are lists, the number of samples
produced by the output function will equal the length of the longer list,
and the shorter list will be repeated until it has that length. In other
words, a separate id and collection of labels for the sides can be
provided for each sample produced by the output function, and if separate
id's or sides are not provided, they will be filled in by repeating
whatever has been provided.
}

