% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_percent}
\alias{make_percent}
\title{Convert a proportion to a percent, leaving a percent unchanged}
\usage{
make_percent(x)
}
\arguments{
\item{x}{The proportion or percent to convert.}
}
\value{
The converted value of \code{x}.
}
\description{
If \eqn{0 \leq x \leq 1}, then this function multiplies \code{x} by 100. If \eqn{1 < x \leq 100}, then
\code{x} is returned unchanged.
}
