% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_proportion}
\alias{make_proportion}
\title{Convert a percent to a proportion, leaving a proportion unchanged}
\usage{
make_proportion(x)
}
\arguments{
\item{x}{The percent or proportion to convert.}
}
\value{
The converted value of \code{x}.
}
\description{
If \eqn{1 < x \leq 100}, then this function divides \code{x} by 100. If  \eqn{0 \leq x \leq 1}, then
\code{x} is returned unchanged.
}
