% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_samplers.R
\name{make_seed}
\alias{make_seed}
\title{Convert a character string to an integer.}
\usage{
make_seed(id)
}
\arguments{
\item{id}{A character vector or a numeric vector.}
}
\value{
This function returns an integer.
}
\description{
\code{make_seed} converts a character string to an integer.
}
\details{
This function converts a character string (the \code{id}) to an integer.
If the inputted charcter string is already an integer, then that integer
is returned unchanged.

The conversion from character string to integer is accomplished by the
\code{digest2int} function in them \code{digest} package.

The main use of this function is to provide a way to construct
finite sampling functions whose probabilities that are unknown to the
user, and which therefore can sensibly be studied using statistical
inference techniques.
}
