% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front_ends.R
\name{roll}
\alias{roll}
\title{Simulate die rolls followed by t means inference}
\usage{
roll(
  n = 1,
  id = "fair",
  N = 1,
  confidence_level = NA,
  null_value = NA,
  true_value = NA,
  significance_level = 0.05,
  alternative = "two-sided",
  sync = NULL,
  sides = 1:6,
  show_summary = TRUE,
  show_plot = TRUE,
  xlim = NA,
  ylim = NA,
  digits = 6
)
}
\arguments{
\item{n}{A vector of one or two positive integers. A single positive integer
specifies the number of simulated rolls of a die per iteration. Two
positive integers specify the numbers of simulated rolls for each of two
dice per iteration.}

\item{id}{The unique one or two id's of the dice whose rolls are to be
simulated. These can be given either as vectors or lists. See the details
below.}

\item{N}{A positive integer specifying the number of iterations to simulate,
each iteration having the number of simulated rolls specified by \code{n}.}

\item{confidence_level}{A number giving the confidence level of the
confidence intervals to be computed.}

\item{null_value}{A number giving the value of the test parameter under the
null hypothesis in the hypothesis tests to be conducted.}

\item{true_value}{A number giving what is to be taken as the true value of
the test parameter, for the purpose of determining which of the
computed confidence intervals contain the true value.}

\item{significance_level}{A number giving the significance level for the
hypothesis tests to be conducted, for the purpose of determining which
of the hypothesis tests attained statistical significance.}

\item{alternative}{A character string specifying which type of alternative
hypothesis should be used: either "two-sided" (or "two.sided"), "less",
or "greater".}

\item{sync}{An integer or character string that is used to set the random
seed immediately prior to the simulations. An integer or \code{NULL} will
be passed to \code{set.seed} directly, while a character string will
first be converted to an integer by \code{\link{make_seed}} and then
passed to \code{set.seed}.}

\item{sides}{This can be given either as a numerical vector, or as a list
of two such numerical vectors. A single numerical vector specifies the
labels for the sides of a die that will be simulated; two such vectors
specify the labels for two dice to be simulated.}

\item{show_summary}{Boolean indicating: should a text summary be printed?}

\item{show_plot}{Boolean indicating: should a plot be displayed?}

\item{xlim}{Numeric vector of length 2 giving the lower and upper limits
of the x axis on the plot. NA means to use the default limits, if any.}

\item{ylim}{Numeric vector of length 2 giving the lower and upper limits
of the y axis on the plot. NA means to use the default limits, if any.}

\item{digits}{Integer giving the number of digits to round values to. A
value of NA indicates no rounding should be done.}
}
\value{
A ggplot2 plot object (whether or not such an object was
displayed).
}
\description{
\code{roll} simulates \code{N} iterations of \code{n} rolls of a die or two
}
