% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_inferers.R
\name{sample_and_infer}
\alias{sample_and_infer}
\title{Generate samples and apply an inference function to them}
\usage{
sample_and_infer(sampler, inferer, n = 1, N = 1, ...)
}
\arguments{
\item{sampler}{Function that returns either an object that can be used as
the first argument of the \code{inferer} function, or else a list whose
elements can be used as the first several arguments of the \code{inferer}
function.}

\item{inferer}{A function whose first arguments (in order) are given by
what is returned by the \code{sampler} function, and which returns an S3
object of class \code{htest}.}

\item{n}{A positive integer or vector of positive integers to be given as
an argument to the \code{sampler} function.}

\item{N}{A positive integer specifying the number of times the
\code{sampler} function is to be called, each call producing an element
in the \code{htest_list} that is returned by \code{sample_and_infer}.}

\item{...}{Any remaining arguments to give to \code{inferer} beyond those
provided by the output of the \code{sampler} function.}
}
\value{
An object of class \code{htest_list}.
}
\description{
Generates samples and applies an inference function to them
}
\details{
This function does the following:

\enumerate{
\item Calls \code{sampler(n)}.
\item Does Item 1 again and again, a total of \code{N} times, resulting in
a list of length \code{N}. This is accomplished by calling
\code{\link{iterate}}.
\item Calls \code{inferer} individually on each element of this list,
with the first argument(s) of \code{inferer} taken from the
element of the list and the remaining argument(s) taken
from \code{...}. This is accomplished by calling
\code{\link{list_infer}}.
}

The result of the third is an S3 object of class \code{hlist}, which is
returned.
}
