% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{t_star}
\alias{t_star}
\title{Compute a \eqn{t} distribution central quantile (\eqn{t^*} value)}
\usage{
t_star(p, df)
}
\arguments{
\item{p}{The level of the central quantile to be computed, expressed either as a proportion (between
0 and 1) or a percent (between 1 and 100).}

\item{df}{The degrees of freedom of the \eqn{t} distribution.}
}
\value{
The central \code{p}-quantile for a \eqn{t} distribution with \code{df} degrees of freedom, which is a number.
}
\description{
This function computes the central \code{p}-quantile for a \eqn{t} distribution with \code{df} degrees of freedom
(which is also called a \eqn{t^*} value).
This is the nonnegative number \eqn{t^*} with the property that the area under the
probability density function of a \eqn{t} distribution with \code{df} degrees of freedom
between \eqn{-t^*} and \eqn{t^*} equals \code{p}.
}
